py/autogen-example/pyproject.toml
[project]
name = "autogen-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.1",
    "autogen-ext[openai]>=0.7.1",
    "python-dotenv>=1.1.1",
]
